Wrong independent variable chose is fixed. Source code of LSTM is moved to cpp files. Tapenade LSTM is added to a tests
