+++ test/cpp/modules/common/LstmTests.cpp
@@ -10,7 +10,8 @@ INSTANTIATE_TEST_CASE_P(Lstm, LstmModuleTest,
     ::testing::Values(
         "../../../../src/cpp/modules/manual/Manual.dll",
         "../../../../src/cpp/modules/manualEigen/ManualEigen.dll",
-        "../../../../src/cpp/modules/finite/Finite.dll"
+        "../../../../src/cpp/modules/finite/Finite.dll",
+		"../../../../src/cpp/modules/tapenade/Tapenade.dll"
     ),
     get_module_name<ModuleTest::ParamType>);
